Time: 96 ms (21.01%), Space: 20.8 MB (49.23%) - LeetHub

+++ 0450-delete-node-in-a-bst/0450-delete-node-in-a-bst.py
@@ -32,7 +32,7 @@ class Solution:
             elif root.right:
                 root.val=findMin(root.right)
                 root.right=self.deleteNode(root.right,root.val)
-            else:
+            elif root.left:
                 root.val=findMax(root.left)
                 root.left=self.deleteNode(root.left,root.val)
